requirements.txt
Flask
flask-socketio
generativeai
torch
transformers
python-dotenv
google-generativeai